Create a custom hook for scrolling horizontally
